Add Slack notifications for error logging in streams
