+++ scripts/cli.ts
@@ -1,11 +1,8 @@
 import { execSync, spawn } from "child_process";
 import fs from "fs";
 import path from "path";
-import {
-  createTestLogger,
-  extractErrorLogs,
-  sendSlackNotification,
-} from "@helpers/logger";
+import { createTestLogger, extractErrorLogs } from "@helpers/logger";
+import { sendSlackNotification } from "@helpers/notifications";
 import "dotenv/config";
 
 interface RetryOptions {
